backend/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.3.0

# Database
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Hunyuan3D-2 dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
einops>=0.6.1
opencv-python>=4.7.0
numpy>=1.21.0
omegaconf>=2.2.0
tqdm>=4.64.0
trimesh>=3.12.0
pymeshlab>=2022.2
pygltflib>=1.15.0
xatlas>=0.0.7
accelerate>=0.20.0
gradio>=3.35.0
rembg>=2.0.50
onnxruntime>=1.15.0

# Additional 3D processing
ninja>=1.11.0
pybind11>=2.10.0

# AI Integration
openai>=1.35.0
tiktoken>=0.7.0

# Background tasks
celery>=5.3.0
flower>=2.0.1

# Utilities
requests>=2.31.0
pillow>=9.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
